+++ modules/apps/export-import/export-import-test/src/testIntegration/java/com/liferay/exportimport/test/ExportedMissingReferenceExportImportTest.java
@@ -45,7 +45,7 @@ import com.liferay.portal.kernel.zip.ZipReader;
 import com.liferay.portal.kernel.zip.ZipReaderFactoryUtil;
 import com.liferay.portal.service.test.ServiceTestUtil;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portlet.PortletBagImpl;
+import com.liferay.portlet.internal.PortletBagImpl;
 import com.liferay.registry.collections.ServiceTrackerCollections;
 import com.liferay.registry.collections.ServiceTrackerList;
