+++ portal-impl/test/integration/com/liferay/portal/security/pacl/test/PortletBagPoolTest.java
@@ -16,7 +16,7 @@ package com.liferay.portal.security.pacl.test;
 
 import com.liferay.portal.kernel.portlet.PortletBagPool;
 import com.liferay.portal.test.rule.PACLTestRule;
-import com.liferay.portlet.PortletBagImpl;
+import com.liferay.portlet.internal.PortletBagImpl;
 
 import org.junit.Assert;
 import org.junit.ClassRule;
